fix: resolve CI failures - Option<MobileLocationData> ScVal trait, test imports, fmt
