+++ contracts/teachlink/src/mobile_platform.rs
@@ -3,9 +3,11 @@
 //! This module implements mobile-optimized features including offline capabilities,
 //! push notifications, and mobile-specific engagement features.
 
-use soroban_sdk::{contracttype, Address, Bytes, Map, Vec, Env, Symbol, symbol_short, panic_with_error};
-use crate::types::*;
 use crate::errors::MobilePlatformError;
+use crate::types::*;
+use soroban_sdk::{
+    contracttype, panic_with_error, symbol_short, Address, Bytes, Env, Map, Symbol, Vec,
+};
 
 const MOBILE_PROFILE: Symbol = symbol_short!("mob_prof");
 const OFFLINE_CONTENT: Symbol = symbol_short!("off_cont");
@@ -23,626 +25,13 @@ const COMPONENT_CONFIG: Symbol = symbol_short!("comp_cfg");
 
 // ========== Mobile Profile Types ==========
 
-
-
-
 // ========== Offline Capabilities ==========
 
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct OfflineSettings {
-    pub auto_download_enabled: bool,
-    pub download_quality: OfflineQuality,
-    pub storage_limit: u64,
-    pub sync_strategy: SyncStrategy,
-    pub offline_duration: u64, // Hours content stays available
-    pub priority_content: Vec<Bytes>,
-    pub compression_enabled: bool,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum OfflineQuality {
-    TextOnly,
-    LowQuality,
-    StandardQuality,
-    HighQuality,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum SyncStrategy {
-    WiFiOnly,
-    WiFiAndCellular,
-    Manual,
-    SmartAdaptive,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct OfflineContent {
-    pub content_id: u64,
-    pub content_type: OfflineContentType,
-    pub local_path: Bytes,
-    pub file_size: u64,
-    pub compressed_size: u64,
-    pub download_date: u64,
-    pub expiry_date: u64,
-    pub is_available: bool,
-    pub version: u32,
-    pub dependencies: Vec<u64>,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum OfflineContentType {
-    VideoLesson,
-    AudioLesson,
-    TextDocument,
-    Quiz,
-    InteractiveExercise,
-    EBook,
-    CourseMaterial,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct SyncQueue {
-    pub pending_uploads: Vec<SyncItem>,
-    pub pending_downloads: Vec<SyncItem>,
-    pub conflict_resolution: Vec<SyncConflict>,
-    pub last_sync_attempt: u64,
-    pub sync_status: SyncStatus,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct SyncItem {
-    pub id: u64,
-    pub item_type: SyncItemType,
-    pub local_path: Bytes,
-    pub remote_path: Bytes,
-    pub priority: SyncPriority,
-    pub retry_count: u32,
-    pub max_retries: u32,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum SyncItemType {
-    ProgressData,
-    QuizResults,
-    Notes,
-    Bookmarks,
-    Certificates,
-    UserPreferences,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum SyncPriority {
-    Low,
-    Normal,
-    High,
-    Critical,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct SyncConflict {
-    pub conflict_id: u64,
-    pub item_type: SyncItemType,
-    pub local_version: Bytes,
-    pub remote_version: Bytes,
-    pub conflict_reason: Bytes,
-    pub resolution_strategy: ConflictResolution,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum ConflictResolution {
-    LocalWins,
-    RemoteWins,
-    Merge,
-    ManualReview,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum SyncStatus {
-    Idle,
-    InProgress,
-    Completed,
-    Failed,
-    Paused,
-}
-
-// ========== Push Notifications ==========
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct NotificationPreferences {
-    pub learning_reminders: bool,
-    pub deadline_alerts: bool,
-    pub achievement_notifications: bool,
-    pub social_updates: bool,
-    pub content_updates: bool,
-    pub quiet_hours: TimeRange,
-    pub frequency_limit: u32, // Max notifications per hour
-    pub sound_enabled: bool,
-    pub vibration_enabled: bool,
-    pub led_enabled: bool,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct TimeRange {
-    pub start_hour: u32,
-    pub end_hour: u32,
-    pub timezone: Bytes,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct PushNotification {
-    pub id: u64,
-    pub user: Address,
-    pub notification_type: NotificationType,
-    pub title: Bytes,
-    pub message: Bytes,
-    pub data: Map<Bytes, Bytes>, // Additional data
-    pub priority: NotificationPriority,
-    pub scheduled_time: u64,
-    pub expiry_time: u64,
-    pub is_read: bool,
-    pub action_buttons: Vec<NotificationAction>,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum NotificationType {
-    LearningReminder,
-    DeadlineAlert,
-    AchievementUnlocked,
-    SocialUpdate,
-    ContentUpdate,
-    SystemMessage,
-    PaymentRequired,
-    CourseUpdate,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum NotificationPriority {
-    Low,
-    Normal,
-    High,
-    Critical,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct NotificationAction {
-    pub action_id: Bytes,
-    pub label: Bytes,
-    pub url: Option<Bytes>,
-    pub auto_dismiss: bool,
-}
-
-// ========== Mobile Payment Integration ==========
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct MobilePaymentMethod {
-    pub id: u64,
-    pub payment_type: PaymentType,
-    pub provider: Bytes,
-    pub account_identifier: Bytes, // Tokenized
-    pub is_default: bool,
-    pub is_verified: bool,
-    pub daily_limit: u64,
-    pub monthly_limit: u64,
-    pub created_at: u64,
-    pub last_used: u64,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum PaymentType {
-    CreditCard,
-    DebitCard,
-    MobileWallet,
-    BankTransfer,
-    Cryptocurrency,
-    CarrierBilling,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct MobileTransaction {
-    pub id: u64,
-    pub user: Address,
-    pub payment_method_id: u64,
-    pub amount: u64,
-    pub currency: Bytes,
-    pub description: Bytes,
-    pub merchant: Bytes,
-    pub status: TransactionStatus,
-    pub timestamp: u64,
-    pub confirmation_code: Bytes,
-    pub fraud_score: u32,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum TransactionStatus {
-    Pending,
-    Completed,
-    Failed,
-    Cancelled,
-    Refunded,
-    Disputed,
-}
-
-// ========== Mobile Security ==========
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct MobileSecuritySettings {
-    pub biometric_enabled: bool,
-    pub biometric_type: BiometricType,
-    pub pin_required: bool,
-    pub two_factor_enabled: bool,
-    pub session_timeout: u32, // Minutes
-    pub encryption_enabled: bool,
-    pub remote_wipe_enabled: bool,
-    pub trusted_devices: Vec<Bytes>,
-    pub login_attempts: u32,
-    pub max_login_attempts: u32,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum BiometricType {
-    Fingerprint,
-    FaceID,
-    Voice,
-    Iris,
-    Pattern,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct SecurityEvent {
-    pub id: u64,
-    pub user: Address,
-    pub event_type: SecurityEventType,
-    pub device_id: Bytes,
-    pub location: Option<MobileLocationData>,
-    pub timestamp: u64,
-    pub severity: SecuritySeverity,
-    pub resolved: bool,
-}
-
-// ========== Mobile Analytics ==========
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct MobileAnalytics {
-    pub user: Address,
-    pub device_analytics: DeviceAnalytics,
-    pub usage_analytics: UsageAnalytics,
-    pub performance_analytics: PerformanceAnalytics,
-    pub engagement_analytics: EngagementAnalytics,
-    pub network_analytics: NetworkAnalytics,
-    pub error_tracking: ErrorTracking,
-    pub last_updated: u64,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct DeviceAnalytics {
-    pub app_version: Bytes,
-    pub os_version: Bytes,
-    pub device_model: Bytes,
-    pub screen_resolution: Bytes,
-    pub memory_usage: u64,
-    pub storage_usage: u64,
-    pub battery_level: u32,
-    pub thermal_state: ThermalState,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum ThermalState {
-    Normal,
-    Warm,
-    Hot,
-    Critical,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct UsageAnalytics {
-    pub session_duration: u32, // Average minutes
-    pub sessions_per_day: u32,
-    pub active_days_per_week: u32,
-    pub peak_usage_hours: Vec<u32>,
-    pub feature_usage: Map<Bytes, u32>,
-    pub screen_time: u64, // Total minutes
-    pub data_consumption: u64, // Bytes
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct PerformanceAnalytics {
-    pub app_load_time: u32, // Milliseconds
-    pub screen_render_time: u32,
-    pub network_latency: u32,
-    pub crash_count: u32,
-    pub anr_count: u32, // Application Not Responding
-    pub memory_leaks: u32,
-    pub battery_drain_rate: u64, // Basis points per hour
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct EngagementAnalytics {
-    pub lesson_completion_rate: u64, // Basis points
-    pub quiz_attempt_rate: u64, // Basis points
-    pub social_interaction_count: u32,
-    pub feedback_submission_rate: u64, // Basis points
-    pub push_notif_response_rate: u64, // Basis points
-    pub feature_adoption_rate: Map<Bytes, u64>, // Basis points
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct NetworkAnalytics {
-    pub connection_type_distribution: Map<NetworkType, u32>,
-    pub average_download_speed: u64, // Kbps
-    pub average_upload_speed: u64,
-    pub connection_stability: u64, // Basis points
-    pub offline_duration: u64, // Minutes per day
-    pub roaming_usage: u64, // Bytes
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct ErrorTracking {
-    pub crash_reports: Vec<CrashReport>,
-    pub anr_reports: Vec<ANRReport>,
-    pub network_errors: Vec<NetworkError>,
-    pub user_reported_issues: Vec<UserIssue>,
-    pub error_rate: u64, // Basis points
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct CrashReport {
-    pub id: u64,
-    pub timestamp: u64,
-    pub app_version: Bytes,
-    pub device_info: Bytes,
-    pub stack_trace: Bytes,
-    pub user_action: Bytes,
-    pub reproducible: bool,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct ANRReport {
-    pub id: u64,
-    pub timestamp: u64,
-    pub duration: u32, // Seconds
-    pub app_state: Bytes,
-    pub device_load: u64, // Basis points
-    pub user_action: Bytes,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct NetworkError {
-    pub id: u64,
-    pub timestamp: u64,
-    pub error_type: NetworkErrorType,
-    pub url: Bytes,
-    pub response_code: u32,
-    pub retry_count: u32,
-    pub resolved: bool,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum NetworkErrorType {
-    Timeout,
-    ConnectionRefused,
-    DNSFailure,
-    SSLHandshakeFailed,
-    ServerError,
-    ClientError,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct UserIssue {
-    pub id: u64,
-    pub timestamp: u64,
-    pub issue_type: Bytes,
-    pub description: Bytes,
-    pub severity: u32,
-    pub user_email: Option<Bytes>,
-    pub resolved: bool,
-    pub resolution: Option<Bytes>,
-}
-
-// ========== Mobile Community Features ==========
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct MobileCommunity {
-    pub user: Address,
-    pub mobile_groups: Vec<MobileGroup>,
-    pub location_sharing: LocationSharingSettings,
-    pub quick_actions: Vec<QuickAction>,
-    pub mobile_challenges: Vec<MobileChallenge>,
-    pub social_features: MobileSocialFeatures,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct MobileGroup {
-    pub id: u64,
-    pub name: Bytes,
-    pub description: Bytes,
-    pub members: Vec<Address>,
-    pub is_location_based: bool,
-    pub meeting_locations: Vec<MobileLocationData>,
-    pub mobile_specific_features: Vec<MobileFeature>,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum MobileFeature {
-    LocationCheckIn,
-    VoiceNotes,
-    PhotoSharing,
-    QuickPolls,
-    EmergencyAlerts,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct LocationSharingSettings {
-    pub enabled: bool,
-    pub sharing_duration: u64, // Hours
-    pub trusted_contacts: Vec<Address>,
-    pub accuracy_level: LocationAccuracy,
-    pub auto_check_in: bool,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum LocationAccuracy {
-    Exact,
-    Approximate,
-    CityLevel,
-    Disabled,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct QuickAction {
-    pub id: u64,
-    pub name: Bytes,
-    pub icon: Bytes,
-    pub action_type: QuickActionType,
-    pub target_screen: Bytes,
-    pub parameters: Map<Bytes, Bytes>,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum QuickActionType {
-    StartLesson,
-    JoinStudyGroup,
-    TakeQuiz,
-    ViewProgress,
-    ContactMentor,
-    ScheduleReminder,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct MobileChallenge {
-    pub id: u64,
-    pub title: Bytes,
-    pub description: Bytes,
-    pub challenge_type: ChallengeType,
-    pub requirements: Vec<Bytes>,
-    pub rewards: ChallengeReward,
-    pub participants: Vec<Address>,
-    pub start_date: u64,
-    pub end_date: u64,
-    pub is_mobile_specific: bool,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum ChallengeType {
-    DailyStreak,
-    WeeklyGoal,
-    SocialLearning,
-    LocationBased,
-    SkillMastery,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct ChallengeReward {
-    pub reward_type: RewardType,
-    pub amount: u64,
-    pub badge: Option<Bytes>,
-    pub certificate: Option<Bytes>,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum RewardType {
-    Points,
-    Badge,
-    Certificate,
-    Discount,
-    PremiumAccess,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct MobileSocialFeatures {
-    pub voice_notes_enabled: bool,
-    pub photo_sharing_enabled: bool,
-    pub location_checkins_enabled: bool,
-    pub quick_polls_enabled: bool,
-    pub emergency_contacts: Vec<Address>,
-    pub study_buddies: Vec<Address>,
-    pub mentor_quick_connect: bool,
-}
-
-// ========== Supporting Types ==========
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct DataUsageTracking {
-    pub total_downloaded: u64,
-    pub total_uploaded: u64,
-    pub cached_data: u64,
-    pub streaming_data: u64,
-    pub last_reset: u64,
-    pub daily_limit: u64,
-    pub warning_threshold: u64, // Basis points
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct MobileAccessibilitySettings {
-    pub screen_reader_enabled: bool,
-    pub high_contrast_enabled: bool,
-    pub large_text_enabled: bool,
-    pub voice_control_enabled: bool,
-    pub gesture_navigation_enabled: bool,
-    pub haptic_feedback_enabled: bool,
-    pub color_blind_mode: ColorBlindMode,
-    pub reduced_motion_enabled: bool,
-    pub focus_indicator_style: FocusStyle,
-}
-
-
+// Types are now imported from crate::types
 
 // ========== Advanced UI/UX Types ==========
 
-
 // ========== Errors ==========
-
 
 // ========== Main Implementation ==========
 
@@ -719,12 +108,12 @@ impl MobilePlatformManager {
                 streaming_data: 0,
                 last_reset: env.ledger().timestamp(),
                 daily_limit: 100 * 1024 * 1024, // 100MB
-                warning_threshold: 8000, // 80% (basis points)
+                warning_threshold: 8000,        // 80% (basis points)
             },
         };
 
         Self::set_mobile_profile(env, &user, &mobile_profile);
-        
+
         Ok(())
     }
 
@@ -738,7 +127,7 @@ impl MobilePlatformManager {
         user.require_auth();
 
         let mut profile = Self::get_mobile_profile(env, &user);
-        
+
         // Check storage availability
         if profile.data_usage.total_downloaded > profile.offline_settings.storage_limit {
             return Err(MobilePlatformError::InsufficientStorage);
@@ -751,19 +140,20 @@ impl MobilePlatformManager {
             file_size: Self::estimate_content_size(content_id, quality.clone()),
             compressed_size: Self::estimate_compressed_size(content_id, quality),
             download_date: env.ledger().timestamp(),
-            expiry_date: env.ledger().timestamp() + profile.offline_settings.offline_duration * 3600,
+            expiry_date: env.ledger().timestamp()
+                + profile.offline_settings.offline_duration * 3600,
             is_available: true,
             version: 1,
             dependencies: Vec::new(env),
         };
 
         Self::add_offline_content(env, &user, &offline_content);
-        
+
         // Update data usage
         profile.data_usage.total_downloaded += offline_content.file_size;
         profile.last_sync = env.ledger().timestamp();
         Self::set_mobile_profile(env, &user, &profile);
-        
+
         Ok(())
     }
 
@@ -792,7 +182,7 @@ impl MobilePlatformManager {
         };
 
         Self::add_push_notification(env, &user, &notification);
-        
+
         Ok(notification_id)
     }
 
@@ -808,7 +198,7 @@ impl MobilePlatformManager {
 
         let profile = Self::get_mobile_profile(env, &user);
         let payment_method = Self::get_payment_method(&profile, payment_method_id);
-        
+
         if payment_method.is_none() {
             return Err(MobilePlatformError::PaymentFailed);
         }
@@ -829,7 +219,7 @@ impl MobilePlatformManager {
         };
 
         Self::add_mobile_transaction(env, &transaction);
-        
+
         Ok(transaction_id)
     }
 
@@ -842,19 +232,33 @@ impl MobilePlatformManager {
         location: Option<MobileLocationData>,
         severity: SecuritySeverity,
     ) -> Result<(), MobilePlatformError> {
+        let (has_location, lat, lon, accuracy, loc_ts) = match location {
+            Some(loc) => (
+                true,
+                loc.latitude,
+                loc.longitude,
+                loc.accuracy,
+                loc.timestamp,
+            ),
+            None => (false, 0i64, 0i64, 0u64, 0u64),
+        };
         let security_event = SecurityEvent {
             id: env.ledger().sequence() as u64,
             user: user.clone(),
             event_type,
             device_id,
-            location,
+            has_location,
+            location_lat: lat,
+            location_lon: lon,
+            location_accuracy: accuracy,
+            location_ts: loc_ts,
             timestamp: env.ledger().timestamp(),
             severity,
             resolved: false,
         };
 
         Self::add_security_event(env, &user, &security_event);
-        
+
         Ok(())
     }
 
@@ -904,7 +308,7 @@ impl MobilePlatformManager {
         }
         status.current_stage = stage;
         status.last_updated = env.ledger().timestamp();
-        
+
         Self::set_onboarding_status(env, &user, &status);
         Ok(())
     }
@@ -927,7 +331,7 @@ impl MobilePlatformManager {
             timestamp: env.ledger().timestamp(),
             category,
         };
-        
+
         Self::add_user_feedback(env, &feedback);
         Ok(feedback_id)
     }
@@ -936,7 +340,7 @@ impl MobilePlatformManager {
     pub fn get_user_experiment_variants(env: &Env, user: Address) -> Map<u64, Symbol> {
         let mut results = Map::new(env);
         let experiments = Self::get_active_experiments(env);
-        
+
         for exp in experiments.iter() {
             if let Some(variant) = exp.variant_allocations.get(user.clone()) {
                 results.set(exp.experiment_id, variant);
@@ -961,9 +365,7 @@ impl MobilePlatformManager {
 
     /// Set design system configuration (admin only)
     pub fn set_design_system_config(env: &Env, config: ComponentConfig) {
-        env.storage()
-            .persistent()
-            .set(&COMPONENT_CONFIG, &config);
+        env.storage().persistent().set(&COMPONENT_CONFIG, &config);
     }
 
     // ========== Helper Functions ==========
@@ -971,10 +373,10 @@ impl MobilePlatformManager {
     fn estimate_content_size(content_id: u64, quality: OfflineQuality) -> u64 {
         // Simulated size estimation
         match quality {
-            OfflineQuality::TextOnly => 1024 * 100, // 100KB
-            OfflineQuality::LowQuality => 1024 * 1024 * 50, // 50MB
+            OfflineQuality::TextOnly => 1024 * 100,               // 100KB
+            OfflineQuality::LowQuality => 1024 * 1024 * 50,       // 50MB
             OfflineQuality::StandardQuality => 1024 * 1024 * 200, // 200MB
-            OfflineQuality::HighQuality => 1024 * 1024 * 500, // 500MB
+            OfflineQuality::HighQuality => 1024 * 1024 * 500,     // 500MB
         }
     }
 
@@ -983,8 +385,14 @@ impl MobilePlatformManager {
         (original_size * 70) / 100 // 30% compression
     }
 
-    fn get_payment_method(profile: &MobileProfile, payment_method_id: u64) -> Option<MobilePaymentMethod> {
-        profile.payment_methods.iter().find(|method| method.id == payment_method_id)
+    fn get_payment_method(
+        profile: &MobileProfile,
+        payment_method_id: u64,
+    ) -> Option<MobilePaymentMethod> {
+        profile
+            .payment_methods
+            .iter()
+            .find(|method| method.id == payment_method_id)
     }
 
     fn generate_confirmation_code(env: &Env) -> Bytes {
@@ -1024,11 +432,9 @@ impl MobilePlatformManager {
             .persistent()
             .get(&key)
             .unwrap_or(Vec::new(env));
-        
+
         contents.push_back(content.clone());
-        env.storage()
-            .persistent()
-            .set(&key, &contents);
+        env.storage().persistent().set(&key, &contents);
     }
 
     fn add_push_notification(env: &Env, user: &Address, notification: &PushNotification) {
@@ -1038,11 +444,9 @@ impl MobilePlatformManager {
             .persistent()
             .get(&key)
             .unwrap_or(Vec::new(env));
-        
+
         notifications.push_back(notification.clone());
-        env.storage()
-            .persistent()
-            .set(&key, &notifications);
+        env.storage().persistent().set(&key, &notifications);
     }
 
     fn add_mobile_transaction(env: &Env, transaction: &MobileTransaction) {
@@ -1052,11 +456,9 @@ impl MobilePlatformManager {
             .persistent()
             .get(&key)
             .unwrap_or(Vec::new(env));
-        
+
         transactions.push_back(transaction.clone());
-        env.storage()
-            .persistent()
-            .set(&key, &transactions);
+        env.storage().persistent().set(&key, &transactions);
     }
 
     fn add_security_event(env: &Env, user: &Address, event: &SecurityEvent) {
@@ -1066,11 +468,9 @@ impl MobilePlatformManager {
             .persistent()
             .get(&key)
             .unwrap_or(Vec::new(env));
-        
+
         events.push_back(event.clone());
-        env.storage()
-            .persistent()
-            .set(&key, &events);
+        env.storage().persistent().set(&key, &events);
     }
 
     fn get_onboarding_status(env: &Env, user: &Address) -> Option<OnboardingStatus> {
@@ -1091,11 +491,9 @@ impl MobilePlatformManager {
             .persistent()
             .get(&USER_FEEDBACK)
             .unwrap_or(Vec::new(env));
-        
+
         feedbacks.push_back(feedback.clone());
-        env.storage()
-            .persistent()
-            .set(&USER_FEEDBACK, &feedbacks);
+        env.storage().persistent().set(&USER_FEEDBACK, &feedbacks);
     }
 
     fn get_active_experiments(env: &Env) -> Vec<UXExperiment> {
